multi_agent_research_system/test_simple_api.py
#!/usr/bin/env python3
"""
Simple test to verify API integration is working.
"""

import asyncio
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Set up API configuration
ANTHROPIC_BASE_URL = os.getenv('ANTHROPIC_BASE_URL')
ANTHROPIC_API_KEY = os.getenv('ANTHROPIC_API_KEY')

if ANTHROPIC_API_KEY:
    os.environ['ANTHROPIC_BASE_URL'] = ANTHROPIC_BASE_URL
    os.environ['ANTHROPIC_API_KEY'] = ANTHROPIC_API_KEY
    print(f"✅ Using Anthropic API: {ANTHROPIC_BASE_URL}")
else:
    print("❌ No API key found")
    exit(1)

from claude_agent_sdk import ClaudeSDKClient, ClaudeAgentOptions


async def test_simple_api():
    """Test simple API connection."""
    print("🔍 Testing API connection...")

    try:
        # Create a simple client
        options = ClaudeAgentOptions(
            model="sonnet",
            max_turns=1
        )

        client = ClaudeSDKClient(options)
        await client.connect()

        # Simple test query
        response = await client.query("Say 'Hello from Claude!'")
        print(f"✅ API Response: {response}")

        # Don't disconnect to avoid the cleanup error for now
        print("✅ API connection test successful!")
        return True

    except Exception as e:
        print(f"❌ API connection failed: {e}")
        return False


async def main():
    """Main test function."""
    print("🚀 Simple API Test")
    print("=" * 30)

    success = await test_simple_api()

    if success:
        print("\n🎉 API INTEGRATION WORKING!")
        print("✅ The system is ready to use with the real Anthropic API")
        print("\n🚀 To run the full web interface:")
        print("   uv run streamlit run ui/streamlit_app.py")
        print("\n📋 The system will now use real Claude responses for:")
        print("   • Research and web searches")
        print("   • Report generation")
        print("   • Quality assessment")
        print("   • Multi-agent coordination")
    else:
        print("\n❌ API integration failed")

    return success


if __name__ == "__main__":
    success = asyncio.run(main())
    exit(0 if success else 1)