multi_agent_research_system/requirements.txt
streamlit>=1.28.0
claude-agent-sdk>=0.1.0
asyncio
pathlib
typing
datetime

# Advanced Scraping Dependencies
crawl4ai==0.7.4
playwright>=1.55.0
pydantic-ai>=1.0.2
cssselect>=1.3.0
beautifulsoup4>=4.12.3
python-dotenv>=1.0.0
httpx>=0.25.0